Move log context data filtering to separate class
