+++ src/Agent/NewRelic/Agent/Core/Agent.cs
@@ -2,10 +2,10 @@
 // SPDX-License-Identifier: Apache-2.0
 
 using NewRelic.Agent.Api;
-using NewRelic.Agent.Api.Experimental;
 using NewRelic.Agent.Configuration;
 using NewRelic.Agent.Core.AgentHealth;
 using NewRelic.Agent.Core.Aggregators;
+using NewRelic.Agent.Core.Attributes;
 using NewRelic.Agent.Core.BrowserMonitoring;
 using NewRelic.Agent.Core.DistributedTracing;
 using NewRelic.Agent.Core.Logging;
@@ -23,14 +23,12 @@ using NewRelic.Agent.Core.Wrapper.AgentWrapperApi.Synthetics;
 using NewRelic.Agent.Extensions.Providers.Wrapper;
 using NewRelic.Core;
 using NewRelic.Core.Logging;
-using NewRelic.SystemExtensions.Collections.Generic;
 using NewRelic.SystemInterfaces;
 using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Text;
-using System.Text.RegularExpressions;
 using System.Threading;
 
 namespace NewRelic.Agent.Core
@@ -62,6 +60,7 @@ namespace NewRelic.Agent.Core
         private readonly ICATSupportabilityMetricCounters _catMetricCounters;
         private readonly Api.ITraceMetadataFactory _traceMetadataFactory;
         private readonly ILogEventAggregator _logEventAggregator;
+        private readonly ILogContextDataFilter _logContextDataFilter;
         private Extensions.Logging.ILogger _logger;
 
         public Agent(ITransactionService transactionService, ITransactionTransformer transactionTransformer,
@@ -71,7 +70,7 @@ namespace NewRelic.Agent.Core
             IBrowserMonitoringPrereqChecker browserMonitoringPrereqChecker, IBrowserMonitoringScriptMaker browserMonitoringScriptMaker,
             IConfigurationService configurationService, IAgentHealthReporter agentHealthReporter, IAgentTimerService agentTimerService,
             IMetricNameService metricNameService, Api.ITraceMetadataFactory traceMetadataFactory, ICATSupportabilityMetricCounters catMetricCounters,
-            ILogEventAggregator logEventAggregator)
+            ILogEventAggregator logEventAggregator, ILogContextDataFilter logContextDataFilter)
         {
             _transactionService = transactionService;
             _transactionTransformer = transactionTransformer;
@@ -91,6 +90,7 @@ namespace NewRelic.Agent.Core
             _traceMetadataFactory = traceMetadataFactory;
             _catMetricCounters = catMetricCounters;
             _logEventAggregator = logEventAggregator;
+            _logContextDataFilter = logContextDataFilter;
 
             Instance = this;
         }
@@ -98,10 +98,6 @@ namespace NewRelic.Agent.Core
         public IConfiguration Configuration => _configurationService.Configuration;
 
         public Extensions.Logging.ILogger Logger => _logger ?? (_logger = new Logger());
-
-        private List<LogContextDataFilterRule> _includeRuleList;
-        private List<LogContextDataFilterRule> _excludeRuleList;
-        private List<LogContextDataFilterRule> _orderedCludeRuleList;
 
         #region Transaction management
 
@@ -451,11 +447,11 @@ namespace NewRelic.Agent.Core
                 {
                     logEventWireModel = new LogEventWireModel(timestamp, logMessage, normalizedLevel,
                         StackTraces.ScrubAndTruncate(logException, LogExceptionStackLimit), logException.Message, logException.GetType().ToString(),
-                        spanId, traceId, FilterLogContextData(logContextData));
+                        spanId, traceId, _logContextDataFilter.FilterLogContextData(logContextData));
                 }
                 else
                 {
-                    logEventWireModel = new LogEventWireModel(timestamp, logMessage, normalizedLevel, spanId, traceId, FilterLogContextData(logContextData));
+                    logEventWireModel = new LogEventWireModel(timestamp, logMessage, normalizedLevel, spanId, traceId, _logContextDataFilter.FilterLogContextData(logContextData));
                 }
 
                 var transaction = _transactionService.GetCurrentInternalTransaction();
@@ -558,91 +554,8 @@ namespace NewRelic.Agent.Core
             _transactionService.RemoveOutstandingInternalTransactions(removeAsync, removePrimary);
         }
 
-        private Dictionary<string, object> FilterLogContextData(Dictionary<string, object> unfilteredContextData)
-        {
-            if (unfilteredContextData == null)
-            {
-                return null;
-            }
-
-            var filteredContextData = new Dictionary<string, object>();
-
-            // Pseudocode
-            // Include/exclude rules:
-            // 1. Exclude wins over include
-            // 2. Can have a '*' wildcard at the end of an include/exclude definition
-            // 3. "more specific" rules win over "less specific".  Example: include=AB, exclude=A*, AB would be included but not ABC or AAA
-
-            // Algorithm:
-            // Create list of rules, sorted by least specific to most specific, with includes first and then excludes within each level of specificity
-            // Specificity is just the length of the string, minus any wildcard character at the end
-            // Apply rules in order. Include rules add attributes to filteredContextData from unfilteredContextData, and exclude rules remove attributes from filteredContextData
-
-            var allRulesInOrder = _orderedCludeRuleList ?? (_orderedCludeRuleList = GetOrderedCludeRuleList()); 
-
-            if (_includeRuleList.Count() == 0) // empty "include" list, so include everything
-            {
-                filteredContextData = unfilteredContextData;
-            }
-
-            // Now apply the rules in order
-            // TODO: maybe optimize by just working with keys as lists of strings until it comes time to return the actual data
-            foreach (var rule in allRulesInOrder)
-            {
-                if (rule.Include)
-                {
-                    filteredContextData = filteredContextData.Concat(unfilteredContextData.Where(x => Regex.IsMatch(x.Key, rule.Text))).ToDictionary();
-                }
-                else
-                {
-                    filteredContextData = filteredContextData.Where(x => !Regex.IsMatch(x.Key, rule.Text)).ToDictionary();
-                }
-            }
-            return filteredContextData;
-        }
-
-        private List<LogContextDataFilterRule> GetCludeRulesFromConfig(bool isIncludeRules)
-        {
-            List<LogContextDataFilterRule> rules = new List<LogContextDataFilterRule>();
-            var configRuleList = isIncludeRules ? _configurationService.Configuration.ContextDataInclude : _configurationService.Configuration.ContextDataExclude;
-            foreach (var configRule in configRuleList)
-            {
-                rules.Add(new LogContextDataFilterRule(configRule, isIncludeRules));
-            }
-            return rules;
-        }
-
-        private List<LogContextDataFilterRule> GetOrderedCludeRuleList()
-        {
-            var includeRuleList = _includeRuleList ?? (_includeRuleList = GetCludeRulesFromConfig(true));
-            var excludeRuleList = _excludeRuleList ?? (_excludeRuleList = GetCludeRulesFromConfig(false));
-
-            // This should work because OrderBy uses a stable sort.  By sorting each list individually,
-            // concactating them, and then sorting again, we get a list of rules ordered by length (specificity),
-            // with includes before excludes for each level of specificity
-            return includeRuleList.OrderBy(rule => rule.Specificity).Concat(
-                excludeRuleList.OrderBy(rule => rule.Specificity))
-                .OrderBy(rule => rule.Specificity).ToList();
-        }
 
         #endregion
     }
 
-    internal class LogContextDataFilterRule
-    {
-        string _text;
-        bool _include;
-        int _specificity; // just the length, except ignore a trailing wildcard.  I.e. AB is more specific than A*
-
-        internal LogContextDataFilterRule(string text, bool include)
-        {
-            _text = text;
-            _include = include;
-            _specificity = text.TrimEnd('*').Length;
-        }
-
-        public string Text => _text;
-        public bool Include => _include;
-        public int Specificity => _specificity;
-    }
 }
